chore: refactored wallet into composable services
