+++ packages/core-web/rollup.config.js
@@ -14,6 +14,6 @@ export default [
       sourcemap: true,
     },
     plugins: [typescript(), terser()],
-    external: ['fedimint-client-wasm'],
+    external: ['@fedimint/fedimint-client-wasm'],
   },
 ]
