+++ packages/core-web/src/transport/index.ts
@@ -1,0 +1,1 @@
+export { WorkerClient } from './WorkerClient'
